+++ lib/cmsg.h
@@ -1,13 +1,13 @@
 #ifndef __CMSG_H__
 #define __CMSG_H__
 
-#include <bits/types.h>
+#include <cdefs.h>
 
 #define SCM_RIGHTS      1
 #define SCM_CREDENTIALS 2
 
 struct cmsg {
-	size_t len;
+	ulong len;
 	int level;
 	int type;
 	char data[];
@@ -19,10 +19,10 @@ struct ucred {
 	int gid;
 };
 
-int cmsg_align(int len);
+ulong cmsg_align(ulong len);
 struct cmsg* cmsg_first(void* p, void* e);
 struct cmsg* cmsg_next(void* p, void* e);
-void* cmsg_put(void* p, void* e, int lvl, int type, void* data, int len);
+void* cmsg_put(void* p, void* e, int lvl, int type, void* data, ulong len);
 
 int cmsg_paylen(struct cmsg* cm);
 void* cmsg_payload(struct cmsg* cm);
